Add ApplyBuffFixer helper: this fabricates `applybuff` events for buffs that were applied pre-combat and drop during the fight Removed the Bloodlust helper since the ApplyBuffFixer now solves this.
